Add support for shiftout / shiftin operations (Issue #12) + Likely fix for Issue #13.

+++ Source/ArduinoDriver/SerialProtocol/AnalogReadResponse.cs
@@ -5,9 +5,9 @@
         public int PinRead { get; private set; }
         public int PinValue { get; private set; }
 
-        public AnalogReadResponse(int pinRead, byte byte1, byte byte2)
+        public AnalogReadResponse(byte pinRead, byte byte1, byte byte2)
         {
-            PinRead = (byte) pinRead;
+            PinRead = pinRead;
             PinValue = (byte1 << 8) + byte2;
         }
     }
